Move SubtensorMixin to types file to avoid import conflicts.
