+++ bittensor/core/types.py
@@ -15,8 +15,206 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
+from abc import ABC
+import argparse
 from typing import TypedDict, Optional
-from bittensor.utils import Certificate
+from bittensor.utils import networking, Certificate
+from bittensor.utils.btlogging import logging
+from bittensor.core import settings
+from bittensor.core.config import Config
+
+
+class SubtensorMixin(ABC):
+    network: str
+    chain_endpoint: str
+    log_verbose: bool
+
+    def __str__(self):
+        return f"Network: {self.network}, Chain: {self.chain_endpoint}"
+
+    def __repr__(self):
+        return self.__str__()
+
+    def _check_and_log_network_settings(self):
+        if self.network == settings.NETWORKS[3]:  # local
+            logging.warning(
+                ":warning: Verify your local subtensor is running on port [blue]9944[/blue]."
+            )
+
+        if (
+            self.network == "finney"
+            or self.chain_endpoint == settings.FINNEY_ENTRYPOINT
+        ) and self.log_verbose:
+            logging.info(
+                f"You are connecting to {self.network} network with endpoint {self.chain_endpoint}."
+            )
+            logging.debug(
+                "We strongly encourage running a local subtensor node whenever possible. "
+                "This increases decentralization and resilience of the network."
+            )
+            # TODO: remove or apply this warning as updated default endpoint?
+            logging.debug(
+                "In a future release, local subtensor will become the default endpoint. "
+                "To get ahead of this change, please run a local subtensor node and point to it."
+            )
+
+    @staticmethod  # TODO can this be a class method?
+    def config() -> "Config":
+        """
+        Creates and returns a Bittensor configuration object.
+
+        Returns:
+            config (bittensor.core.config.Config): A Bittensor configuration object configured with arguments added by
+                the `subtensor.add_args` method.
+        """
+        parser = argparse.ArgumentParser()
+        SubtensorMixin.add_args(parser)
+        return Config(parser)
+
+    @staticmethod
+    def setup_config(network: Optional[str], config: "Config"):
+        """
+        Sets up and returns the configuration for the Subtensor network and endpoint.
+
+        This method determines the appropriate network and chain endpoint based on the provided network string or
+            configuration object. It evaluates the network and endpoint in the following order of precedence:
+            1. Provided network string.
+            2. Configured chain endpoint in the `config` object.
+            3. Configured network in the `config` object.
+            4. Default chain endpoint.
+            5. Default network.
+
+        Arguments:
+            network (Optional[str]): The name of the Subtensor network. If None, the network and endpoint will be
+                determined from the `config` object.
+            config (bittensor.core.config.Config): The configuration object containing the network and chain endpoint
+                settings.
+
+        Returns:
+            tuple: A tuple containing the formatted WebSocket endpoint URL and the evaluated network name.
+        """
+        if network is None:
+            candidates = [
+                (
+                    config.is_set("subtensor.chain_endpoint"),
+                    config.subtensor.chain_endpoint,
+                ),
+                (config.is_set("subtensor.network"), config.subtensor.network),
+                (
+                    config.subtensor.get("chain_endpoint"),
+                    config.subtensor.chain_endpoint,
+                ),
+                (config.subtensor.get("network"), config.subtensor.network),
+            ]
+            for check, config_network in candidates:
+                if check:
+                    network = config_network
+
+        evaluated_network, evaluated_endpoint = (
+            SubtensorMixin.determine_chain_endpoint_and_network(network)
+        )
+
+        return networking.get_formatted_ws_endpoint_url(
+            evaluated_endpoint
+        ), evaluated_network
+
+    @classmethod
+    def help(cls):
+        """Print help to stdout."""
+        parser = argparse.ArgumentParser()
+        cls.add_args(parser)
+        print(cls.__new__.__doc__)
+        parser.print_help()
+
+    @classmethod
+    def add_args(cls, parser: "argparse.ArgumentParser", prefix: Optional[str] = None):
+        """
+        Adds command-line arguments to the provided ArgumentParser for configuring the Subtensor settings.
+
+        Arguments:
+            parser (argparse.ArgumentParser): The ArgumentParser object to which the Subtensor arguments will be added.
+            prefix (Optional[str]): An optional prefix for the argument names. If provided, the prefix is prepended to
+                each argument name.
+
+        Arguments added:
+            --subtensor.network: The Subtensor network flag. Possible values are 'finney', 'test', 'archive', and
+                'local'. Overrides the chain endpoint if set.
+            --subtensor.chain_endpoint: The Subtensor chain endpoint flag. If set, it overrides the network flag.
+            --subtensor._mock: If true, uses a mocked connection to the chain.
+
+        Example:
+            parser = argparse.ArgumentParser()
+            Subtensor.add_args(parser)
+        """
+        prefix_str = "" if prefix is None else f"{prefix}."
+        try:
+            default_network = settings.DEFAULT_NETWORK
+            default_chain_endpoint = settings.FINNEY_ENTRYPOINT
+
+            parser.add_argument(
+                f"--{prefix_str}subtensor.network",
+                default=default_network,
+                type=str,
+                help="""The subtensor network flag. The likely choices are:
+                                        -- finney (main network)
+                                        -- test (test network)
+                                        -- archive (archive network +300 blocks)
+                                        -- local (local running network)
+                                    If this option is set it overloads subtensor.chain_endpoint with
+                                    an entry point node from that network.
+                                    """,
+            )
+            parser.add_argument(
+                f"--{prefix_str}subtensor.chain_endpoint",
+                default=default_chain_endpoint,
+                type=str,
+                help="""The subtensor endpoint flag. If set, overrides the --network flag.""",
+            )
+            parser.add_argument(
+                f"--{prefix_str}subtensor._mock",
+                default=False,
+                type=bool,
+                help="""If true, uses a mocked connection to the chain.""",
+            )
+
+        except argparse.ArgumentError:
+            # re-parsing arguments.
+            pass
+
+    @staticmethod
+    def determine_chain_endpoint_and_network(
+        network: str,
+    ) -> tuple[Optional[str], Optional[str]]:
+        """Determines the chain endpoint and network from the passed network or chain_endpoint.
+
+        Arguments:
+            network (str): The network flag. The choices are: ``finney`` (main network), ``archive`` (archive network
+                +300 blocks), ``local`` (local running network), ``test`` (test network).
+
+        Returns:
+            tuple[Optional[str], Optional[str]]: The network and chain endpoint flag. If passed, overrides the
+                ``network`` argument.
+        """
+
+        if network is None:
+            return None, None
+        if network in settings.NETWORKS:
+            return network, settings.NETWORK_MAP[network]
+
+        substrings_map = {
+            "entrypoint-finney.opentensor.ai": ("finney", settings.FINNEY_ENTRYPOINT),
+            "test.finney.opentensor.ai": ("test", settings.FINNEY_TEST_ENTRYPOINT),
+            "archive.chain.opentensor.ai": ("archive", settings.ARCHIVE_ENTRYPOINT),
+            "subvortex": ("subvortex", settings.SUBVORTEX_ENTRYPOINT),
+            "127.0.0.1": ("local", settings.LOCAL_ENTRYPOINT),
+            "localhost": ("local", settings.LOCAL_ENTRYPOINT),
+        }
+
+        for substring, result in substrings_map.items():
+            if substring in network:
+                return result
+
+        return "unknown", network
 
 
 class AxonServeCallParams(TypedDict):
